requirements.txt
torch==2.1.0
transformers==4.34.0
sqlean.py
python-dotenv
mistralai